+++ app/public/js/form-validators/emailValidator.js
@@ -1,8 +1,7 @@
 
-function EmailValidator(){
-
+function EmailValidator()
+{
 // bind this to _local for anonymous functions //
-
 	var _local = this;
 
 // modal window to allow users to request credentials by email //
@@ -10,7 +9,6 @@ function EmailValidator(){
 	_local.retrievePassword.modal({ show : false, keyboard : true, backdrop : true });
 	_local.retrievePasswordAlert = $('#get-credentials .alert');
 	_local.retrievePassword.on('show', function(){ $('#get-credentials-form').resetForm(); _local.retrievePasswordAlert.hide();});
-
 }
 
 EmailValidator.prototype.validateEmail = function(e)