+++ app/public/js/views/login.js
@@ -1,6 +1,6 @@
 
 $(document).ready(function(){
-	
+
 	var lv = new LoginValidator();
 	var lc = new LoginController();
 
@@ -48,4 +48,4 @@ $(document).ready(function(){
 		}
 	});
 	
-})
+});